Refactoring: Fixed NPE in SettingsManager, extracted CamelCatalogService, and simplified class hierarchy
