+++ designite-output/DesigniteLog29112025_2120.txt
@@ -1,0 +1,6 @@
+29-11-2025 21:20:36: Detecting code smells...
+29-11-2025 21:20:36: Detecting cyclic dependency...
+29-11-2025 21:20:36: Detecting architecture smells...
+29-11-2025 21:20:37: Exporting analysis results...
+29-11-2025 21:20:45: Done.
+Total time (sec): 17
